Creating a separate reducer for the todos
