+++ store/reducers/index.js
@@ -1,55 +1,6 @@
-import {
-  CREATE_TODO,
-  DELETE_TODO,
-  CHANGE_CURRENT_NAME,
-  CHANGE_CURRENT_DESCRIPTION,
-} from "../actions/";
+import { combineReducers } from "redux";
+import todo from "./todoReducer";
 
-let initialState = {
-  todos: [],
-  todo: {
-    name: "",
-    description: "",
-    id: 0,
-  },
-};
-
-const rootReducer = (state = initialState, action) => {
-  switch (action.type) {
-    case CREATE_TODO:
-      if (state.todo.name !== "" && state.todo.description !== "") {
-        return {
-          ...state,
-          todos: [
-            ...state.todos,
-            {
-              ...state.todo,
-              id: state.todos.length + 1,
-            },
-          ],
-        };
-      } else return state;
-
-    case DELETE_TODO:
-      return {
-        ...state,
-        todos: state.todos.filter((todo) => todo.id !== action.payload),
-      };
-    case CHANGE_CURRENT_NAME:
-      let nameChange = { ...state.todo, name: action.payload };
-      return {
-        ...state,
-        todo: nameChange,
-      };
-    case CHANGE_CURRENT_DESCRIPTION:
-      let newTodo = { ...state.todo, description: action.payload };
-      return {
-        ...state,
-        todo: { ...newTodo },
-      };
-    default:
-      return state;
-  }
-};
-
-export default rootReducer;
+export default combineReducers({
+  todo,
+});
